+++ evalml/pipelines/components/estimators/regressors/__init__.py
@@ -6,4 +6,4 @@ from .xgboost_regressor import XGBoostRegressor
 from .et_regressor import ExtraTreesRegressor
 from .baseline_regressor import BaselineRegressor
 from .decision_tree_regressor import DecisionTreeRegressor
-from .timeseries_baseline_regressor import TimeSeriesBaselineRegressor
+from .time_series_baseline_regressor import TimeSeriesBaselineRegressor
